+++ .gitignore
@@ -260,3 +260,4 @@ project.json
 # Archive files
 *.zip
 /build/dotnet-script
+/dotnet-script
