+++ src/Dotnet.Script.Core/Interactive/ClsCommand.cs
@@ -1,0 +1,12 @@
+﻿namespace Dotnet.Script.Core
+{
+    public class ClsCommand : IInteractiveCommand
+    {
+        public string Name => "cls";
+
+        public void Execute(CommandContext commandContext)
+        {
+            commandContext.Console.Clear();
+        }
+    }
+}
