+++ src/Dotnet.Script.Core/Interactive/CommandContext.cs
@@ -1,0 +1,14 @@
+﻿namespace Dotnet.Script.Core
+{
+    public class CommandContext
+    {
+        public CommandContext(ScriptConsole console, InteractiveRunner runner)
+        {
+            Console = console;
+            Runner = runner;
+        }
+
+        public ScriptConsole Console { get; }
+        public InteractiveRunner Runner { get; }
+    }
+}
