+++ src/Dotnet.Script.Core/Interactive/IInteractiveCommand.cs
@@ -1,0 +1,8 @@
+﻿namespace Dotnet.Script.Core
+{
+    public interface IInteractiveCommand
+    {
+        string Name { get; }
+        void Execute(CommandContext commandContext);
+    }
+}
