+++ src/Dotnet.Script.Core/Interactive/InteractiveRunner.cs
@@ -12,82 +12,6 @@ using System;
 
 namespace Dotnet.Script.Core
 {
-    public class CommandContext
-    {
-        public CommandContext(ScriptConsole console, InteractiveRunner runner)
-        {
-            Console = console;
-            Runner = runner;
-        }
-
-        public ScriptConsole Console { get; }
-        public InteractiveRunner Runner { get; }
-    }
-
-    public class ResetInteractiveCommand : IInteractiveCommand
-    {
-        public string Name => "reset";
-
-        public void Execute(CommandContext commandContext)
-        {
-            commandContext.Runner.Reset();
-        }
-    }
-
-    public class ClsCommand : IInteractiveCommand
-    {
-        public string Name => "cls";
-
-        public void Execute(CommandContext commandContext)
-        {
-            commandContext.Console.Clear();
-        }
-    }
-
-    public class ExitCommand : IInteractiveCommand
-    {
-        public string Name => "exit";
-
-        public void Execute(CommandContext commandContext)
-        {
-            Environment.Exit(0);
-        }
-    }
-
-    public interface IInteractiveCommand
-    {
-        string Name { get; }
-        void Execute(CommandContext commandContext);
-    }
-
-    public class InteractiveCommandProvider
-    {
-        private IInteractiveCommand[] _commands = new IInteractiveCommand[] 
-        {
-            new ResetInteractiveCommand(),
-            new ClsCommand(),
-            new ExitCommand()
-        };
-
-        public bool TryProvideCommand(string code, out IInteractiveCommand command)
-        {
-            command = null;
-
-            // not a REPL command
-            if (!code.StartsWith("#")) return false;
-
-            // compiler level directive
-            if (code.StartsWith("#r ") || code.StartsWith("#load ")) return false;
-
-            var commandName = code?.Split(' ')[0]?.Substring(1)?.Trim();
-            if (commandName == null) return false;
-
-            command = _commands.FirstOrDefault(x => x.Name == commandName);
-
-            return command != null;
-        }
-    }
-
     public class InteractiveRunner
     {
         private ScriptState<object> _scriptState;
