+++ src/Dotnet.Script.Core/Interactive/ResetInteractiveCommand.cs
@@ -1,0 +1,12 @@
+﻿namespace Dotnet.Script.Core
+{
+    public class ResetInteractiveCommand : IInteractiveCommand
+    {
+        public string Name => "reset";
+
+        public void Execute(CommandContext commandContext)
+        {
+            commandContext.Runner.Reset();
+        }
+    }
+}
